Imported method to JobsFragment
